add utlis.py
